backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-docx
playwright
python-multipart
